refactor: move storage implementations to their own namespace
