--- Doppler.HtmlEditorApi/Infrastructure/IDatabaseConnectionFactory.cs
@@ -1,10 +1,0 @@
-using System.Data;
-using System.Threading.Tasks;
-
-namespace Doppler.HtmlEditorApi.Infrastructure
-{
-    public interface IDatabaseConnectionFactory
-    {
-        Task<IDbConnection> GetConnection();
-    }
-}
