+++ Doppler.HtmlEditorApi/Storage.DapperProvider/IDatabaseConnectionFactory.cs
@@ -1,0 +1,9 @@
+using System.Data;
+using System.Threading.Tasks;
+
+namespace Doppler.HtmlEditorApi.Storage.DapperProvider;
+
+public interface IDatabaseConnectionFactory
+{
+    Task<IDbConnection> GetConnection();
+}
